+++ server/src/main/java/org/opensearch/search/internal/SearchContext.java
@@ -560,6 +560,11 @@ public abstract class SearchContext implements Releasable {
         return 0;
     }
 
+    @ExperimentalApi
+    public long termsAggregationMaxPrecomputeCardinality() {
+        return 30_000L;
+    }
+
     public CardinalityAggregationContext cardinalityAggregationContext() {
         return new CardinalityAggregationContext(false, Runtime.getRuntime().maxMemory() / 100);
     }